PL/3110/code-backup-20251031_202116/backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# MQTT
paho-mqtt==1.6.1

# Redis
redis==5.0.1
hiredis==2.2.3

# WebSocket
websockets==12.0

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2
